CONFIG Modified application_stop.sh

+++ scripts/application_stop.sh
@@ -1,4 +1,15 @@
 #!/bin/bash
-#Stopping existing node servers
+
+# Stopping existing node servers
 echo "Stopping any existing node servers"
-pkill node
+pidof node
+RESULT=$?
+
+if test $RESULT -eq 0
+then
+    echo "Found node process, stopping node"
+else
+    echo "No node process found"
+    pkill node
+fi
+echo $?
